refactor: HeaderのeditNameSectionのロジックの分離
