+++ frontend/src/components/header/hooks/useEditImageSection.tsx
@@ -1,0 +1,3 @@
+export const useEditImageSection = () => {
+
+}
